Moved text and testbed classes
Moved text and testbed classes to separate files TextClasses.py and Testbed.py
Also moved a bunch of constants that went with the classes.
